BUILD : 1.7.6 (518) BUGS : 262 NOTES : CS ACCESS was also affected by the overflow
git-svn-id: svn://svn.anope.org/anope/trunk@518 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@372 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/chanserv.c
@@ -4028,7 +4028,7 @@ static int do_access(User * u)
     ChanAccess *access;
 
     int i;
-    short level = 0, ulev;
+    int level = 0, ulev;
     int is_list = (cmd && stricmp(cmd, "LIST") == 0);
     int is_servadmin = is_services_admin(u);
